// src/generated/sources/ais.ts
export interface SecuritySale {
  /** "SFT-17-LES(M)" */
  infoCode: string;
  /** Description */
  infoDescription: string;
  /** e.g., CDSL, NSDL */
  informationSource: string;
  /** Trade date */
  dateOfSaleTransfer: string;
  /** Name of the stock / unit */
  securityName: string;
  /** ISIN code */
  securityCodeIsin: string;
  /** e.g., "Listed Equity Share" */
  securityClass: string;
  /** Trade channel */
  debitType: string;
  /** Settlement type */
  creditType: string;
  /** "Short term" or "Long term" */
  assetType: string;
  /** Quantity sold */
  quantity: number;
  /** Price unit */
  salePricePerUnit: number;
  /** Gross value */
  salesConsideration: number;
  /** Buying cost */
  costOfAcquisition: number;
  /** Unit Fair Market Value */
  unitFmv: number;
  /** Absolute FMV */
  fairMarketValue: number;
  /** Indexed buying cost */
  indexedCostOfAcquisition: number;
  /** Transaction status */
  status: string;
}